Build-tool related changes so the Haxe runtime types are generated into the main game module

+++ InitPlugin.hx
@@ -13,14 +13,9 @@ class InitPlugin {
     trace("Game location: " + target);
     var gameDir = Path.directory(target);
     var pluginPath = Sys.getCwd();
-    // compile the Haxe build plugin
-    trace("Building BuildTool...");
-    var cmd = Sys.command('haxe',['--cwd','$pluginPath/Haxe/BuildTool','build.hxml']);
-    if (cmd != 0) throw "Haxe BuildTool compilation failed";
 
-    updateProject(gameDir, pluginPath, new Path(target).file);
-
-    trace("Project update done. Don't forget to add the 'HaxeRuntime' project to your .Target.cs file!");
+    updateProject(gameDir, pluginPath, new Path(target).file, true);
+    trace("Project update done.");
   }
 
   static function inferProjectLocation() {
@@ -34,12 +29,26 @@ class InitPlugin {
     throw 'Cannot find uproject location! Try to specify it as an argument';
   }
 
-  public static function updateProject(gameDir:String, pluginPath:String, projectName:String)
-  {
+  static function getHaxeModule(gameDir:String) {
+    var dirs = [];
+    for (file in FileSystem.readDirectory('$gameDir/Source')) {
+      if (file != 'HaxeRuntime' && FileSystem.isDirectory('$gameDir/Source/$file'))
+        dirs.push(file);
+    }
+    if (dirs.length == 1)
+      return dirs[0];
+    else
+      throw 'Multiple modules not supported'; //TODO
+  }
+
+  public static function updateProject(gameDir:String, pluginPath:String, projectName:String, fromCommandLine=false) {
+    gameDir = FileSystem.fullPath(gameDir);
+    pluginPath = FileSystem.fullPath(pluginPath);
+
     trace('Updating game project...');
     updateGameProject(gameDir, projectName);
     trace('Updating game module...');
-    updateGameModule(gameDir, pluginPath);
+    updateGameModule(gameDir, pluginPath, fromCommandLine);
     inline function checkDir(dir:String) {
       if (!FileSystem.exists(dir))
         FileSystem.createDirectory(dir);
@@ -49,8 +58,9 @@ class InitPlugin {
     checkDir('$gameDir/Haxe/Externs');
   }
 
-  private static function updateGameModule(gameDir:String, pluginPath:String)
+  private static function updateGameModule(gameDir:String, pluginPath:String, fromCommandLine:Bool)
   {
+    var mod = getHaxeModule(gameDir);
     // update templates that need to be updated
     function recurse(templatePath:String, toPath:String, alsoDelete:Bool)
     {
@@ -63,6 +73,9 @@ class InitPlugin {
 
       for (file in FileSystem.readDirectory(templatePath))
       {
+        // Only copy the module-specific code if
+        if ( mod != 'HaxeRuntime' && (file == 'HaxeRuntime.cpp' || file == 'HaxeRuntime.h') )
+          continue;
         if (checkMap != null) checkMap[file] = true;
         var curTemplPath = '$templatePath/$file',
             curToPath = '$toPath/$file';
@@ -72,7 +85,9 @@ class InitPlugin {
         } else {
           var shouldCopy = !FileSystem.exists(curToPath) || file.endsWith('.cs');
           var contents = File.getContent(curTemplPath);
-          if (!shouldCopy)
+          if (mod != 'HaxeRuntime')
+            contents = contents.replace('HAXERUNTIME', mod.toUpperCase()).replace('HaxeRuntime', mod);
+          if (!shouldCopy && file != 'arguments.hxml')
             shouldCopy = contents != File.getContent(curToPath);
 
           if (shouldCopy)
@@ -87,21 +102,65 @@ class InitPlugin {
             deleteRecursive('$toPath/$file');
       }
     }
-    recurse('$pluginPath/Haxe/Templates/Source', '$gameDir/Source', false);
+
+    recurse('$pluginPath/Haxe/Templates/Source/HaxeRuntime', '$gameDir/Source/$mod', false);
+    recurse('$pluginPath/Haxe/Templates/Haxe', '$gameDir/Haxe', false);
+    // TODO: take this off once we can decide where the plugin dir will be
+    if (FileSystem.exists('$gameDir/Source/HaxeRuntime'))
+      deleteRecursive('$gameDir/Source/HaxeRuntime', true);
+
+    handleModuleRules(gameDir, pluginPath, mod, fromCommandLine);
   }
 
-  private static function deleteRecursive(path:String)
+  private static function handleModuleRules(gameDir:String, pluginPath:String, module:String, alsoCompile:Bool) {
+    if (!FileSystem.exists('$gameDir/Source/$module/$module.Build.hx')) {
+      throw 'ERROR: $module.Build.hx does not exist. Please port your old $module.Build.cs to Haxe and re-run init-plugin';
+    }
+    gameDir = gameDir.replace('\\','/');
+    pluginPath = pluginPath.replace('\\','/');
+    var args = [
+      '# use this to build the build tool. if you have made any changes to any .Build.hx files, run this',
+      '-cp $pluginPath/Haxe/BuildTool/src',
+      '-cp $pluginPath/Haxe/BuildTool',
+      '-cp $pluginPath',
+      '-net-lib $pluginPath/Haxe/BuildTool/lib/UnrealBuildTool.dll',
+      '-cs $gameDir/Intermediate/Haxe/BuildTool',
+      'HaxeInit',
+      '',
+      '-D no-root',
+      '-D net_ver=40',
+      '-D analyzer',
+      '-D real_position',
+      '--macro Package.main("$gameDir/Source", [{ name:"HaxeInit", target:"$pluginPath/Source/HaxeInit/HaxeInit.Build.cs" }])'
+    ];
+    sys.io.File.saveContent('$gameDir/Haxe/build-module-rules.hxml', args.join('\n'));
+
+    if (alsoCompile) {
+      // compile the Haxe build plugin
+      trace("Building BuildTool...");
+      var cmd = Sys.command('haxe',['$gameDir/Haxe/build-module-rules.hxml']);
+      if (cmd != 0) throw "Haxe BuildTool compilation failed";
+    }
+  }
+
+  private static function deleteRecursive(path:String, force=false):Bool
   {
+    var shouldDelete = true;
     if (!FileSystem.isDirectory(path))
     {
       FileSystem.deleteFile(path);
     } else {
       for (file in FileSystem.readDirectory(path)) {
-        if (file != 'Generated')
-          deleteRecursive('$path/$file');
+        if (force || (file != 'Private' && file != 'Public')) {
+          shouldDelete = deleteRecursive('$path/$file',force);
+        } else {
+          shouldDelete = false;
+        }
       }
-      FileSystem.deleteDirectory(path);
+      if (shouldDelete)
+        FileSystem.deleteDirectory(path);
     }
+    return shouldDelete;
   }
 
   private static function updateGameProject(gameDir:String, projectName:String)
@@ -109,13 +168,19 @@ class InitPlugin {
     var projFile = gameDir + '/$projectName.uproject';
     var props = haxe.Json.parse(File.getContent(projFile));
     var modules:Array<{ Name:String, Type:String, LoadingPhase:String }> = props.Modules;
+    // TODO take this off once we support multiple modules
+    var allMods = [], shouldChange = false;
     for (module in modules)
     {
       if (module.Name == 'HaxeRuntime')
-        return; //already there
+        shouldChange = true;
+      else
+        allMods.push(module);
     }
 
-    modules.push({ Name:'HaxeRuntime', Type:'Runtime', LoadingPhase:'Default' });
-    File.saveContent(projFile, haxe.Json.stringify(props));
+    if (shouldChange) {
+      props.Modules = allMods;
+      File.saveContent(projFile, haxe.Json.stringify(props));
+    }
   }
 }
